requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
neo4j>=5.17.0
python-dotenv>=1.0.0
